PostgreSQLConnection -> ServerConnection
